# Optimize incrementing values in DirectFileStore adapter