@dmagliola 2023-02-28
I'm not opposed to this change, but:
- I'm not sure it actually makes a measurable performance difference, two out of 3 benchmarks report no change. Note that in my experience, testing performance of the disk store is very hard to get meaningful results. We used to use a cloudvm to get realistic / repeatable results, testing locally was all over the place. 
- sounds like what's you're trying to avoid is the double call to internal_store. You could avoid that by assigning the result of internal_store to a local var, then you don't need the new method in FileMappedDict.
- avoiding the double lookup in the positions hash should be below measurability. Especially when compared to writing to the file...