@splattael 2023-03-01
@dmagliola Thanks a ton for reviewing and merging :bow:

You are right regarding benchmarks - they were flaky for me locally too (on Linux) but in the end they gave me overall speed improvement. This improvement makes sense (not only because of side-stepping `Process.pid`) but also having less calls.

BTW, I am planning to reduce the amount of `getpid` (via `Process.pid`) syscalls by setting the PID on startup and override `Kernel.fork` (or `Process._fork` in Ruby 3.1+). See https://bugs.ruby-lang.org/issues/5446 and https://bugs.ruby-lang.org/issues/17795 for more context.

:wave: @SuperQ :heart: