@splattael 2023-03-01
> In the meantime, i'd recommend doing just the change to sidestep the double call to internal_store, which adds almost zero complexity, and doesn't hurt, really :)

I had this very refactoring ready in another branch but saw less speed improvements. :sweat_smile: 

I should have provided more context before, sorry about :bow: