@dmagliola 2023-03-04
No worries, and thanks for the PR!
I'm going to wait on @Sinjo to confirm he's cool with this, but this should go out in the next release :)

> I am planning to reduce the amount of getpid (via Process.pid) syscalls by setting the PID on startup and override Kernel.fork (or Process._fork in Ruby 3.1+).

I may be misreading [the docs](https://ruby-doc.org/3.1.3/Process.html) here, but is this a method intended to be monkeypatched? 😅 

That makes me a bit uncomfortable, but it seems like it's designed precisely to do that...
It's unfortunate the underscore in the name makes it hard to google for examples, [all I can find is the test case where it gets introduced](https://github.com/ruby/ruby/pull/5017/files#diff-c7bbc51b9f06aa0c63621b1a074a0befed88ec432ba375d88e6f365f635d1d0eR2625-R2631)...

So I guess we could do that `prepend` in the constructor of `DirectFileStore`.

Would that look like:

```
def _fork
  super
  DirectFileStore.process_pid = Process.pid 
end
```

Am I understanding that correctly?