@splattael 2023-03-05
@dmagliola Yes, is the way I understood it as well :+1: 

It seems that Rails, for example, already does exactly this for Ruby 3.1+. See https://github.com/rails/rails/blob/v7.0.2/activesupport/lib/active_support/fork_tracker.rb#L50