@dmagliola 2023-03-05
Ah, that's amazing! You found exactly the kind of example I was hoping to get! 🙌 
That sounds like a sensible thing to do, and i'm looking forward to the PR!

NOTE: I would do "half" of what Rails is doing. I'd prepend if `_fork` is defined, but if not (if Ruby < 3.1), i'd leave it alone and continue checking for `Process.pid` on every increment. The logic here being that `_fork` is intended for you to do this, but the other 3 aren't. I'm open minded about this. The argument "look, Rails does it and here's exactly how to do it safely" is pretty compelling, I just think we should limit how much we pollute the global namespace with "surprising" stuff.