@splattael 2023-03-07
@dmagliola @Sinjo I agree! We should use `_fork` automatically if available :+1: 

I wonder, though, if we could provide a way for users to opt-in to use cached process ids (via :monkey: patching `Process.fork`)? :thinking: 

So:
* For Ruby 3.1+ - cached process id on by default via `_fork` hook
* For <Ruby 3.1 - off by default but opt-in allowed if setup code is called explicitly

WDYT?